This is the current build for XOmB Bare Bones.
Since GDC is for all reasonable purposes dead in development, LDC is the current way to go. In this build, LDC is used to compile any D language code. You will not need a naked-compile of GDC to compile anymore.

We will no longer maintain GDC as a first priority.

Please check the wiki first for build instructions, check the README for a link.